Break down components.
